[controller] Extract parent push job lifecycle logic into ParentPushJobLifecycleManager
